requirements.txt
aniso8601==8.0.0
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
blinker==1.4
click==7.1.2
Deprecated==1.2.9
dnspython==1.16.0
email-validator==1.1.0
Flask==1.1.2
Flask-BabelEx==0.9.4
flask-buzz==0.1.15
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-praetorian==1.0.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pendulum==2.1.0
pluggy==0.13.1
py==1.8.1
py-buzz==1.0.3
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pytzdata==2019.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
